feat: generate weight CSS files directly and major refactor
