+++ package.json
@@ -30,7 +30,7 @@
     "format:scripts": "prettier --write ./scripts/**/*.{js,jsx,json,md}",
     "format:eslint": "eslint **/*.{js,jsx} --fix",
     "format:prettier": "prettier --write **/*.{js,jsx,json,md}",
-    "build:generic": "node scripts/generic/generic-font-packager.js",
+    "build:generic": "node scripts/generic/build.js",
     "build:google": "node scripts/google/download-google.js",
     "build:google-force": "node scripts/google/download-google.js force",
     "parser:v1": "npm explore google-font-metadata -- yarn parser:v1",
